Update message.spec.js
updated line 27 from message.value or something to message.commands; forgot I changed that. Back to tests passing but npm test still fails. No real test for test 6 exists currently, which is why it passes. I need to make sure an array is being passed in as the 2nd argument in the constructor for test 6. Time to go back to the Jasmine Documentation